# perf(es/transform): Add inline annotation