--- a/Cargo.toml
+++ b/Cargo.toml
@@ -150,7 +150,8 @@ resolver = "2"
   wasmer-wasix              = { version = "0.27.0", default-features = false }
 
 [profile.release]
-lto = true
+debug = true
+# lto = true
 
 # We use CARGO_PROFILE_RELEASE_LTO for production builds
 # lto = "fat"
@@ -160,7 +161,7 @@ lto = true
 
 [profile.bench]
 debug = true
-lto   = true
+# lto   = true
 
 # Optimize for iteration
 [profile.dev.build-override]