… crates/swc/benches/assets/Observable.ts