**Description:**

### swc_core@v3.0.1, LTO = false

```
Benchmarking es/full/all/es5: Collecting 100 samples in estimated 54.860 s (100 iterations)
Benchmarking es/full/all/es5: Analyzing
es/full/all/es5         time:   [525.54 ms 526.92 ms 528.33 ms]
```

### swc_core@v3.0.1, LTO = false + `#[inline(always)]`

```
Benchmarking es/full/all/es5: Collecting 100 samples in estimated 53.857 s (100 iterations)
Benchmarking es/full/all/es5: Analyzing
es/full/all/es5         time:   [510.40 ms 512.59 ms 514.88 ms]
```


**Related issue (if exists):**